requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.21.13
botocore==1.24.13
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.11
django-environ==0.8.1
django-import-export==2.7.1
django-mathfilters==1.0.0
django-storages==1.12.3
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
